requirements.txt
werkzeug
fastapi
uvicorn
python-multipart
jinja2
ultralytics
torch
torchvision
numpy
Pillow
